requirements.txt
awscli==1.11.104
boto3==1.4.5
botocore==1.5.67
kubernetes==3.0.0
PyYAML==3.12
Jinja2==2.9.6
docker==2.5.1
